+++ packages/ui/src/cairo/GovernorControls.svelte
@@ -242,4 +242,4 @@
 
 <MacrosSection bind:macros={opts.macros} />
 
-<InfoSection bind:info={opts.info} />
+<InfoSection bind:info={opts.info} {errors} />