+++ packages/ui/src/cairo_alpha/ERC721Controls.svelte
@@ -149,4 +149,4 @@
 
 <MacrosSection bind:macros={opts.macros} />
 
-<InfoSection bind:info={opts.info} />
+<InfoSection bind:info={opts.info} {errors} />