+++ packages/ui/src/solidity/AccountControls.svelte
@@ -232,4 +232,4 @@
   disabledReason={upgradeNotSupportedReason}
 />
 
-<InfoSection bind:info={opts.info} />
+<InfoSection bind:info={opts.info} {errors} />