+++ packages/ui/src/stellar/GovernorControls.svelte
@@ -104,4 +104,4 @@
 
 <TraitImplementationSection bind:explicitImplementations={opts.explicitImplementations} />
 
-<InfoSection bind:info={opts.info} />
+<InfoSection bind:info={opts.info} {errors} />